+++ src/core/cli.rs
@@ -285,7 +285,7 @@ pub fn run_headless(do_niseci: bool, args: &Vec<String>) -> bool {
                 elenco_specie: riferimento_specie
             };
             match calculate_niseci(&campionamento, &riferimento, &anagrafica) {
-                Ok(niseci) => {
+                Ok((niseci, _intermediates)) => {
                     let rqe_niseci = calculate_rqe_niseci(niseci);
                     let stato_eco_niseci = calculate_stato_ecologico(niseci, &anagrafica.area);
                     println!("NISECI: {niseci}");